requirements.txt
transformers
pytorch
python_dotenv